Splitted xrange series into type files.
